requirements.txt
wheel==0.44.0
dronekit==2.9.2
dronekit_sitl
pyserial==3.5
GitPython==3.1.42
pytest==8.0.1
halo~=0.0.31
selenium==4.17.2
requests==2.31.0
PyGithub==2.2.0
pymsteams==0.2.2
git+ssh://git@github.com/spearuav/ninox_common@1.43.2
git+ssh://git@github.com/spearuav/ninox-python-infra-lib@1.16.5
webdriver_manager==4.0.1
prettytable==3.9.0
packaging==23.2
colorama==0.4.6
console-menu==0.8.0
ansicolors==1.1.8
paramiko==3.4.0
docker==6.1.3
pexpect==4.9.0
ConsoleMenu==1.0.1
monday==2.0.0rc3
mavproxy==1.8.71
pyttsx3==2.98
